requirements.txt
RPi.GPIO
alsaaudio
vlc
pydbus
